Refactored Candidates
